Implement hunk relevance end to end
